Adding notes to database
Added in correct API logic for receiving a new note and adding it to the correct project in the database. This ensures notes are added to the correct project, adjusted render logic for notes list to use projects held in redux state. This makes the list rerender correctly when a new note is added and the projects list is fetched again.